Laska.jl documentation

Summarizing statistics

Calculating frequency

Helpers

ISI

Laska.isi — Method

isi(cluster::Cluster{T}) where {T<:Real}

Calculate the ISI of each spike in cluster.

Spike filtering

By depth

Laska.spikesatdepth — Method

spikesatdepth(experiment::E{T}, depth::N) where {E<:AbstractExperiment,T<:Real,N<:Real}
spikesatdepth(experiment::E{T}, depth::Tuple{2,N}) where {E<:AbstractExperiment,T<:Real,N<:Real}
spikesatdepth(experiment::E{T}, depth::Set{N}) where {E<:AbstractExperiment,T<:Real,N<:Real}

Returns a Vector{T} of all spiketimes at/in depth.

The included depths are controlled by the type of the depth variable:

A single number returns only the spikes of clusters at that exact depth.

A Tuple with 2 entries returns all clusters at depths between (and including) the values.

A Set returns the clusters with the exact depths in the Set.

Rounding

roundup()

Laska.roundup — Method

roundup(value::T, interval::N) where T<:Real where N<:Real

Rounds value up to the nearest greater multiple of interval.

Examples

julia> Laska.roundup(12, 30)30
julia> Laska.roundup(12, 23)23
julia> Laska.roundup(12.5, 1)13
julia> Laska.roundup(12, 3.8)15.2