Helpers

This section describes general helper functions that does not fit under other, more specific headings.

Helpers

Rounding

Functions for "rounding" numbers to arbitrary intervals.

LaskaCore.roundup — Function

roundup(value::T, interval::N) where T<:Real where N<:Real

Rounds value up to the nearest greater multiple of interval.

Examples

julia> LaskaCore.roundup(12, 30)30
julia> LaskaCore.roundup(12, 23)23
julia> LaskaCore.roundup(12.5, 1)13
julia> LaskaCore.roundup(17, 3.8)19.0

LaskaCore.rounddown — Function

rounddown(value::T, interval::N) where {T<:Real} where {N<:Real}

Rounds value down to the nearest lesser multiple of interval.

Examples

julia> LaskaCore.rounddown(12, 30)0
julia> LaskaCore.rounddown(12, 23)0
julia> LaskaCore.rounddown(12.5, 1)12
julia> LaskaCore.rounddown(17, 3.8)15.2

LaskaCore.arbitraryround — Function

arbitraryround(value::T, interval::N) where {T<:Real} where {N<:Real}

Rounds value to the nearest multiple of interval.

Examples

julia> LaskaCore.arbitraryround(12, 30)0
julia> LaskaCore.arbitraryround(12, 23)23
julia> LaskaCore.arbitraryround(12.5, 1)12
julia> LaskaCore.arbitraryround(17, 3.8)15.2

Unit conversions

Functions for converting between different units.

Time

LaskaCore.mstosamplerate
LaskaCore.sampleratetoms