Modelling

Hodgkin-Huxley model

Model containers

LaskaML.hh.HHModel — Type

struct HHModel{T<:Number, U<:Number}
    channels::Vector{HHChannel{T, U}}

    function HHModel(channels::HHChannel{T,U}...) where {T<:Number, U<:Number}
           new{T, U}([channels...])
    end
end

Struct for holding several HHChannels, representing an (almost complete) model.

The channels are indexable using model[i].

The struct can be initialized by calling its inner constructor and passing it the channels to be part of the model. For convenience, the struct may also be initialized by "adding" HHChannels together using the + operator. The + operator may also be used to add HHChannels to an existing HHModel

Examples

Ion channels

Generalized ion channel struct

LaskaML.hh.HHChannel — Type

struct HHChannel{T <: Number, U <: Number}
    id::String
    g::T
    Vsteady::T
    alpha_m::Function
    beta_m::Function
    p::U
    alpha_h::Function
    beta_h::Function
    q::U
end

General struct for a channel in the Hodgkin-Huxley model in the form:

\[I_x = \bar{g}_xm^p_xh^q_x(V_mV_x)\]

The gating variables $m$ and $h$ are are given by α_m, β_m, α_h and β_h respectively in the following wa

\[\begin{aligned} \frac{dm}{dt} = \alpha(V_M)(1-m) - \beta(V_M)m \\ \frac{dh}{dt} = \alpha(V_M)(1-h) - \beta(V_M)h \end{aligned}\]

LaskaML.hh.hhchannel — Function

hhchannel(
id,
g::T,
Vsteady::T;
alpha_m=() -> oneunit(T),
beta_m=() -> oneunit(T),
p=zero(T),
alpha_h=() -> oneunit(T),
beta_h=() -> oneunit(T),
q=zero(T)
) where T<:Number

Outer constructor of HHChannel. If not specified, exponents of $m$ and $h$ will be 0 while unspecified α and β functions will be initialized as () -> oneunit(T).

Interface functions

Functions for accessing the fields of HHChannel

LaskaML.hh.id — Function

id(channel::HHChannel)

Get the identifier of a HHChannel.

LaskaML.hh.conductance — Function

conductance(channel::HHChannel)

Get the maximum conductance, $\bar{g}$, of a HHChannel.

LaskaML.hh.alpha_m — Function

alpha_m(channel::HHChannel) = channel.α_m

Get the function for $lpha_m$ bound to a HHChannel.

LaskaML.hh.beta_m — Function

beta_m(channel::HHChannel)

Get the function for $eta_m$ bound to a HHChannel.

LaskaML.hh.m_exponent — Function

m_exponent(channel::HHChannel)

Get the exponent of $m$ in a HHChannel

LaskaML.hh.alpha_h — Function

alpha_h(channel::HHChannel)

Get the function for $lpha_h$ bound to a HHChannel.

LaskaML.hh.beta_h — Function

beta_h(channel::HHChannel)

Get the function for $eta_h$ bound to a HHChannel.

LaskaML.hh.h_exponent — Function

h_exponent(channel::HHChannel)

Get the exponent of $h$ in a HHChannel.

Expression constructors

The following functions are used to turn HHChannel structs into Strings representing their mathematical expression.

LaskaML.hh.parsechannel — Function

parsechannel(channel::HHChannel)

Parse a HHChannel struct into an expression in the form of a String suitable for use in DifferentialEquations.

Examples

julia> chanK = LaskaML.hh.hhchannel(
               "K",
               35.0,
               -77.0,
               alpha_m=alpha_n,
               beta_m=beta_n,
               p=4,
               alpha_h=alpha_h,
               beta_h=beta_h,
               q=0
           )
LaskaML.hh.HHChannel{Float64, Int64}("K", 35.0, -77.0, alpha_n, beta_n, 4, alpha_h, beta_h, 0)

julia> LaskaML.hh.parsechannel(chanK)
"(gK*(m_K^4)*(v-E_K))"

α and β functions

Generalized version

LaskaML.hh.ab_generalized — Function

ab_generalized(V, p)

The generalized form of α and β.

function ab_generalized(V, p)
    A, B, C, D, F, H = p
    (A + B * V) / (C + H * exp((V + D) / F))
end

A generalized version of the α/β functions as per Nelson, M.E. (2004)

Calculated as:

\[\alpha(V) = \frac{A + BV}{C+Hexp(\frac{V+D}{F})}\]