Hodgkin Huxley models

LaskaML.hh.HHChannel — Type

struct HHChannel{G<:Number, V<:Number, P<:Number, Q<:Number}
    id::String
    g::G  # Max conductance
    Vrev::V # Ion specific reversal potential
    alpha_m::Union{Function, Nothing} # Alpha for gating variable m
    beta_m::Union{Function, Nothing} # Beta for gating variable m
    p::P # Exponent of gating variable m
    alpha_h::Union{Function, Nothing} # Alpha for gating variable h
    beta_h::Union{Function, Nothing} # Beta for garing variable h
    q::Q # Exponent for gating variable h
    m_inf::Union{Function, Nothing}
    h_inf::Union{Function, Nothing}
end

General struct for a channel in the Hodgkin-Huxley model in the form:

\[I_x = \bar{g}_xm^p_xh^q_x(V_m-V_x)\]

The gating variables $m$ and $h$ are are given by α_m, β_m, α_h and β_h respectively in the following wa

\[\begin{aligned} \frac{dm}{dt} = \alpha(V_M)(1-m) - \beta(V_M)m \\ \frac{dh}{dt} = \alpha(V_M)(1-h) - \beta(V_M)h \end{aligned}\]

LaskaML.hh.HHModel — Type

struct HHModel{G<:Number, V<:Number, P<:Number, Q<:Number}
    V_0::G
    I_0::G
    C_0::G
    channels::Vector{HHChannel{G,V,P,Q}}

    function HHModel(V_0, I_0, C_0, channels::HHChannel{G, V, P, Q}...) where {G<:Number, V<:Number, P<:Number, Q<:Number}
        new{G,V,P,Q}(V_0, I_0, C_0, [channels...])
    end
end

Struct for holding several HHChannels, representing an (almost complete) model.

The channels are indexable using model[i].

The struct can be initialized by calling its inner constructor and passing it the channels to be part of the model. For convenience, the struct may also be initialized by "adding" HHChannels together using the + operator. The + operator may also be used to add HHChannels to an existing HHModel

LaskaML.hh.Vrev — Method

 Vrev(channel::HHChannel) = channel.Vsteady

Get the reversal potential.

LaskaML.hh.ab_generalized — Method

ab_generalized(V, p)

The generalized form of α and β.

function ab_generalized(V, p)
    A, B, C, D, F, H = p
    (A + B * V) / (C + H * exp((V + D) / F))
end

LaskaML.hh.alpha_h — Method

alpha_h(channel::HHChannel)

Get the function for $lpha_h$ bound to a HHChannel.

LaskaML.hh.alpha_m — Method

alpha_m(channel::HHChannel) = channel.α_m

Get the function for $lpha_m$ bound to a HHChannel.

LaskaML.hh.beta_h — Method

beta_h(channel::HHChannel)

Get the function for $eta_h$ bound to a HHChannel.

LaskaML.hh.beta_m — Method

beta_m(channel::HHChannel)

Get the function for $eta_m$ bound to a HHChannel.

LaskaML.hh.buildd0 — Method

buildd0(model::HHModel{G,V,P,Q}) where {G<:Number, V<:Number, P<:Number, Q<:Number}

Construct the initial conditions as well as the functions for $frac{dm}{dv}$ and/or $frac{dh}{dv}$ for each channel.

Returns a Tuple consisting of:

A Vector of the functions mentioned above.

A Vector of the initial conditions.

A Vector containing Strings identifying the functions.

LaskaML.hh.buildexpression — Method

buildexpression(model::HHModel{G,V,P,Q}) where {G<:Number, V<:Number, P<:Number, Q<:Number}

Build a string representation of the expression:

\[I = \bar{g}_xm_x^ph_x^q(V_M-V_x) + \bar{g}_ym_y^ph_y^q(V_M-V_y) + \bar{g}_zm_z^ph_z^q(V_M-V_z)...\]

Using all HHChannels in the model. If the exponents $p$ or $q$ of a HHChannel is zero their respective term ($m$ or $h$) will not be included.

Examples

julia> # Example potassium channel
           chanK = LaskaML.hh.hhchannel(
               "K",
               35.0,
               -77.0,
               alpha_m=alpha_n,
               beta_m=beta_n,
               p=4,
               alpha_h=alpha_h,
               beta_h=beta_h,
               q=0
           )
LaskaML.hh.HHChannel{Float64, Float64, Int64, Int64}("K", 35.0, -77.0, alpha_n, beta_n, 4, alpha_h, beta_h, 0)

julia> # Example sodium channel
           chanNa = LaskaML.hh.hhchannel(
               "Na",
               40.0,
               55.0,
               alpha_m=alpha_m,
               beta_m=beta_m,
               p=3,
               alpha_h=alpha_h,
               beta_h=beta_h,
               q=1
           )
LaskaML.hh.HHChannel{Float64, Float64, Int64, Int64}("Na", 40.0, 55.0, alpha_m, beta_m, 3, alpha_h, beta_h, 1)

julia> chanL = LaskaML.hh.hhchannel(
               "L",
               0.3,
               -65.0
           )
LaskaML.hh.HHChannel{Float64, Float64, Int64, Int64}("L", 0.3, -65.0, nothing, nothing, 0, nothing, nothing, 0)

julia> model = chanNa + chanK + chanL # Combine channels into a HHModel
LaskaML.hh.HHModel{Float64, Float64, Int64, Int64}(LaskaML.hh.HHChannel{Float64, Float64, Int64, Int64}[LaskaML.hh.HHChannel{Float64, Float64, Int64, Int64}("Na", 40.0, 55.0, alpha_m, beta_m, 3, alpha_h, beta_h, 1), LaskaML.hh.HHChannel{Float64, Float64, Int64, Int64}("K", 35.0, -77.0, alpha_n, beta_n, 4, alpha_h, beta_h, 0), LaskaML.hh.HHChannel{Float64, Float64, Int64, Int64}("L", 0.3, -65.0, nothing, nothing, 0, nothing, nothing, 0)])
nel{Float64, Float64, Int64, Int64}("L", 0.3, -65.0, nothing, nothing, 0, nothing, nothing, 0)

julia> LaskaML.hh.buildexpression(model) # Build String expression from model.
"(g_Na*(m_Na^3)*h_Na*(v-E_Na))+(g_K*(m_K^4)*(v-E_K))+(g_L*(v-E_L))"

LaskaML.hh.buildmodel — Method

buildmodel(model::HHModel{G,V,P,Q}) where {G<:Number, V<:Number, P<:Number, Q<:Number}

Returns a Tuple containing all parts necessary for constructing an ODEProblem of the model except the time span:

A mutating function accepting the parameters du, u, p, t

The initial conditions of the model

The parameters of the model

LaskaML.hh.buildparams — Method

buildparams(model::HHModel{G,V,P,Q}) where {G<:Number, V<:Number, P<:Number, Q<:Number}

returns the initial parameters of the model as well as a Vector{String} of their identifiers.

LaskaML.hh.buildterm — Method

function buildterm(variable::String, id::String, exponent::T) where T<:Number

Build a String representation of $m^p$ or $h^q$. If the exponent of either term is 0 it is not included.

LaskaML.hh.conductance — Method

conductance(channel::HHChannel)

Get the maximum conductance, $\bar{g}$, of a HHChannel.

LaskaML.hh.dhdv — Method

 function dhdv(channel::HHChannel)

Returns a function giving dh/dv according to:

\[\frac{dh}{dv} = \alpha(v)(1.0 - h) - \beta(v)h\]

LaskaML.hh.dmdv — Method

 function dmdv(channel::HHChannel)

Returns a function giving dm/dv according to:

\[\frac{dm}{dv} = \alpha(v)(1.0 - m) - \beta(v)m\]

LaskaML.hh.h_exponent — Method

h_exponent(channel::HHChannel)

Get the exponent of $h$ in a HHChannel.

LaskaML.hh.h_inf — Method

h_inf(channel::HHChannel)

Get the function for steady state of $h$.

LaskaML.hh.hhchannel — Method

hhchannel(
id,
g::T,
Vsteady::T;
alpha_m=() -> oneunit(T),
beta_m=() -> oneunit(T),
p=zero(T),
alpha_h=() -> oneunit(T),
beta_h=() -> oneunit(T),
q=zero(T)
) where T<:Number

Outer constructor of HHChannel. If not specified, exponents of $m$ and $h$ will be 0 while unspecified α and β functions will be initialized as nothing.

LaskaML.hh.id — Method

id(channel::HHChannel)

Get the identifier of a HHChannel.

LaskaML.hh.m_exponent — Method

m_exponent(channel::HHChannel)

Get the exponent of $m$ in a HHChannel

LaskaML.hh.m_inf — Method

m_inf(channel::HHChannel)

Get the function for steady state of $m$.

LaskaML.hh.parsechannel — Method

parsechannel(channel::HHChannel)

Parse a HHChannel struct into an expression in the form of a String suitable for use in DifferentialEquations.

Examples

julia> chanK = LaskaML.hh.hhchannel(
               "K",
               35.0,
               -77.0,
               alpha_m=alpha_n,
               beta_m=beta_n,
               p=4,
               alpha_h=alpha_h,
               beta_h=beta_h,
               q=0
           )
LaskaML.hh.HHChannel{Float64, Int64}("K", 35.0, -77.0, alpha_n, beta_n, 4, alpha_h, beta_h, 0)

julia> LaskaML.hh.parsechannel(chanK)
"(gK*(m_K^4)*(v-E_K))"