Import data

Phy output data

LaskaCore.importphy — Function

importphy(phydir::String, glxdir::String, triggerpath::String; includemua::Bool=false)
importphy(phydir::String, filters::Tuple{Symbol,Function}, glxdir::String="", triggerpath::String=""; includemua::Bool=false)
importphy(phydir::String, filters::Tuple{Tuple{Symbol,Function}}, glxdir::String="", triggerpath::String=""; includemua::Bool=false)

Import Kilosort output processed in Phy. Spiketimes are sorted.

By default, only "good" clusters as per phy output are included. Setting includemua=true will include "mua" clusters as well as unclassified. Clusters may be further filtered based on any variable in "cluster_info.tsv". This is done by including a Tuple with the column to be filtered as a symbol and a filtering function. Several Tuples containing such filters may be included by wrapping them in a Tuple.

Example

# Removes any cluster with a mean firing rate less than 1:

# A function returning true if x > 1
function filterfunc(x::Float64)
    return x > 1
end

# A tuple with the above function (filterfunc) and the column to which it should be
# applied (:fr).
filtertuple = (:fr, filterfunc)

result = importphy("phyoutput_directory", "glxoutput_directory", "direct_path_to_triggerfile", filtertuple)

SpikeGLX

Functions for importing and parsing spikeGLX binary and meta files.

LaskaCore.spikemmap — Function

spikemmap(file::String, nchans::Int, filesizebytes::Int)
spikemmap(file::String, meta::Dict{Substring{String}, SubString{String}})

Create a memory map of a spikeGLX .bin file. Requires a path to the file, the number of channels and the file size in bytes. The easiest way to provide this is to pass a parsed .meta file.

LaskaCore.parseglxmeta — Function

parseglxmeta(file::String)

Parses a spikeGLX .meta file into a Dict.

LaskaCore.tovolts — Function

tovolts(in::Matrix{Int16}, meta::Dict{SubString{String},SubString{String}})

Convert a Vector/Matrix of raw spikeGLX Int16 data to volts.

LaskaCore.importchanint16bin — Function

importchanint16bin(path::String)

Import a .bin file containing a SINGLE channel.

LaskaCore.importchanint16csv — Function

importchanint16csv(path::String)

Import a .csv file containing a SINGLE channel.

LaskaCore.gettrig — Function

gettrig(t::Vector{T}) where {T<:Real}

Extract trigger event indices.

Assumes that t is 0 except at trigger events. Returns only the first index at each trigger event(ie series of sequential nonzero indices).

Example

# Example "triggerchannel"
# * = Indices that will be returned, ^ = trigger
#

#            *                 *
v = [0,0,0,0,1,1,1,1,0,0,0,0,0,1,1,1,1,1,1]
#            ^ ^ ^ ^           ^ ^ ^ ^ ^ ^ 

LaskaCore.gettrig(v)