Types and their interfaces

Abstract types

AbstractExperiment

Laska.AbstractExperiment — Type

abstract type AbstractExperiment{T} end

Parent type to concrete types representing entire experiments containing their specifications, metadata and clusters.

Interfaces

The following functions provide interfaces to all structs that are children of Laska.AbstractExperiment.

Laska.clusterids — Function

clusterids(experiment::T) where {T<:AbstractExperiment}

Returns a Vector of all cluster id:s present in experiment.

Laska.getcluster — Function

function getcluster(experiment::T, cluster::Int64) where {T<:AbstractExperiment}

Returns a cluster from experiment.

Laska.clustervector — Function

clustervector(experiment::T) where {T<:AbstractExperiment}

Returns a Vector{T} where T<:AbstractCluster containing all clusters in experiment.

Laska.getmeta — Function

getmeta(experiment::T, entry::String) where {T<:AbstractExperiment}
getmeta(experiment::T) where {T<:AbstractExperiment}

Returns experiment meta info from spikeGLX. If an entry string is not supplied all entries are returned.

Laska.triggertimes — Function

triggertimes(experiment::T) where {T<:AbstractExperiment}

Returns the timestamps of trigger events in experiment.

Laska.ntrigs — Function

ntrigs(experiment::T) where {T<:AbstractExperiment}

Returns the number of trigger events in experiment.

AbstractCluster

Laska.AbstractCluster — Type

abstract type AbstractCluster{T} end

Parent type to concrete types representing single clusters.

Interfaces

The following functions provide interfaces to all structs that are children of Laska.AbstractCluster.

Laska.id — Function

id(cluster::Cluster)

Returns the id of cluster

Laska.nspikes — Function

nspikes(cluster::T) where {T<:AbstractCluster}

Returns the number of spikes (length of the spiketimes field) of Cluster.

Laska.info — Function

info(cluster::Cluster)

Returns info (as dict) about cluster. A string may be supplied to return a specific entry (as Float64).

Laska.spiketimes — Function

spiketimes(cluster::Cluster)

Returns the spiketimes of cluster.

Concrete types

"Experiment" wrappers

PhyOutput

Laska.PhyOutput — Type

mutable struct PhyOutput{T} <: AbstractExperiment{T}
    clusterids::Vector{Int64}
    clusters::Vector{Cluster{T}}
    trigtimes::Vector{T}
    meta::Dict{SubString{String},SubString{String}}
end

Struct for holding Kilosort output preprocessed in Phy. Should be instantiated using the outer constructor Laska.importphy.

Direct field access is not recommended. Basic interface functions include:

Laska.clusterids – Returns all cluster ID:s as a Vector.

Laska.getcluster – Returns a specific Laska.Cluster.

Laska.clustervector – Returns all Laska.Cluster:s.

Laska.getmeta – Returns the spikeGLX meta as a dict or a specific entry.

Laska.triggertimes – Returns the trigger event times.

Laska.ntrigs – Returns the length of the trigger event time Vector.

Single cluster wrappers

Cluster

Laska.Cluster — Type

struct Cluster{T} <: AbstractCluster{T}
    id::Int64
    info::Dict{String,String}
    spiketimes::Vector{T}
end

Struct for holding a single Cluster.

Direct field access is not recommended. Basic interface functions include:

Laska.id – Returns the Cluster id.

Laska.nspikes – Returns the number of spikes in the cluster (Based off length of the spiketimes field).

Laska.info – Returns the info of the Cluster from "cluster_info.tsv" as a Dict.

Laska.spiketimes – Returns a Vector containing all spiketimes of the Cluster.