Laska.jl documentation

Summarizing statistics

Calculating CV2

Laska.cv2 — Method

cv2(cluster::Cluster)

Returns CV2 values of cluster as a vector.

CV2 is calculated according to:

\[CV2 = \frac{2|ISI_{n+1} - ISI_n|}{(ISI_{n+1} + ISI_n)}\]

Laska.cv2mean — Method

cv2mean(cluster::Cluster)

Calculates the mean CV2 of cluster.

Calculating frequency

Laska.frequency — Method

frequency(cluster::Cluster, period::T) where {T<:Real}

Returns a Vector containing the frequency of the cluster in the form of spikes/period binned at each multiple of period. Spiketimes are binned to the next largest multiple of period. Ie a spike happening at time = 30001 will be in the 60000 bin.

Example

For a cluster sampled at 30 000Hz...

Laska.frequency(cluster, 30000)

...will return spikes/second.

Indexing into the result as:

result[n]

Will return the n:th bin which describes the number of spikes occuring between period * n and period * n-1.

Helpers

ISI

Laska.isi — Method

isi(cluster::Cluster{T}) where {T<:Real}

Calculate the ISI of each spike in cluster.

Spike filtering

By depth

Laska.spikesatdepth — Method

spikesatdepth(p::PhyOutput{T}, depth::N) where {T<:Real} where {N<:Real}
spikesatdepth(p::PhyOutput{T}, depth::Tuple{2,N}) where {T<:Real} where {N<:Real}
spikesatdepth(p::PhyOutput{T}, depth::Set{N}) where {T<:Real} where {N<:Real}

Returns a Vector{T} of all spiketimes at/in depth.

The included depths are controlled by the type of the depth variable:

A single number returns only the spikes of clusters at that exact depth.

A Tuple with 2 entries returns all clusters at depths between (and including) the values.

relativefrequency(vec::Vector{Vector{T}}, period::N) where {T<:Real,N<:Real}

A Set returns the clusters with the exact depths in the Set.

Rounding

roundup()

Laska.roundup — Method

roundup(value::T, interval::N) where T<:Real where N<:Real

Rounds value up to the nearest greater multiple of interval.

Examples

julia> Laska.roundup(12, 30)30
julia> Laska.roundup(12, 23)23
julia> Laska.roundup(12.5, 1)13
julia> Laska.roundup(12, 3.8)15.2