Structs and their Interfaces

Abstract types

AbstractExperiment

LaskaCore.AbstractExperiment — Type

abstract type AbstractExperiment{T} end

Parent type to concrete types representing entire experiments containing their specifications, metadata and clusters.

Interfaces

The following functions provide interfaces to all structs that are children of LaskaCore.AbstractExperiment.

LaskaCore.clusterids — Function

clusterids(experiment::T) where {T<:AbstractExperiment}

Returns a Vector of all cluster id:s present in experiment.

LaskaCore.getcluster — Function

function getcluster(experiment::T, cluster::Int64) where {T<:AbstractExperiment}

Returns a cluster from experiment.

LaskaCore.clustervector — Function

clustervector(experiment::T) where {T<:AbstractExperiment}

Returns a Vector{T} where T<:AbstractCluster containing all clusters in experiment.

LaskaCore.getmeta — Function

getmeta(experiment::T, entry::String) where {T<:AbstractExperiment}
getmeta(experiment::T) where {T<:AbstractExperiment}

Returns experiment meta info from spikeGLX. If an entry string is not supplied all entries are returned.

LaskaCore.triggertimes — Function

triggertimes(experiment::T) where {T<:AbstractExperiment}

Returns the timestamps of trigger events in experiment.

LaskaCore.ntrigs — Function

ntrigs(experiment::T) where {T<:AbstractExperiment}

Returns the number of trigger events in experiment.

AbstractCluster

LaskaCore.AbstractCluster — Type

abstract type AbstractCluster{T} end

Parent type to concrete types representing single clusters.

Interfaces

The following functions provide interfaces to all structs that are children of Laska.AbstractCluster.

LaskaCore.id — Function

function id(cluster::T) where {T<:AbstractCluster}

Returns the id of cluster

LaskaCore.nspikes — Function

nspikes(cluster::T) where {T<:AbstractCluster}

Returns the number of spikes (length of the spiketimes field) of Cluster.

LaskaCore.info — Function

info(cluster::T) where {T<:AbstractCluster}

Returns info (as dict) about cluster. A string may be supplied to return a specific entry (as Float64).

LaskaCore.spiketimes — Function

spiketimes(cluster::Cluster::T) where {T<:AbstractCluster}

Returns the spiketimes of cluster.

Concrete types

"Experiment" wrappers

This section describes concrete types for holding entire experiments including their clusters and metadata.

PhyOutput

LaskaCore.PhyOutput — Type

mutable struct PhyOutput{T} <: AbstractExperiment{T}
    clusterids::Vector{Int64}
    clusters::Vector{Cluster{T}}
    trigtimes::Vector{T}
    meta::Dict{SubString{String},SubString{String}}
end

Struct for holding Kilosort output preprocessed in Phy. Should be instantiated using the outer constructor LaskaCore.importphy.

Direct field access is not recommended. Basic interface functions include:

LaskaCore.clusterids – Returns all cluster ID:s as a Vector.

LaskaCore.getcluster – Returns a specific LaskaCore.Cluster.

LaskaCore.clustervector – Returns all LaskaCore.Cluster:s.

LaskaCore.getmeta – Returns the spikeGLX meta as a dict or a specific entry.

LaskaCore.triggertimes – Returns the trigger event times.

LaskaCore.ntrigs – Returns the length of the trigger event time Vector.

RelativeSpikes

LaskaCore.RelativeSpikes — Type

mutable struct RelativeSpikes{T} <: AbstractExperiment{T}
    clusterids::Vector{Int64}
    clusters::Vector{RelativeCluster{T}}
    trigtimes::Vector{T}
    meta::Dict{SubString{String},SubString{String}}
    stimtrain::Dict{String,T}
    specs::Dict{String,T}
end

Similar to LaskaCore.PhyOutput. However, instead of LaskaCore.Clusters, LaskaCore.RelativeClusters are used. In these, spiketimes relative to trigger event(s) instead of absolute spiketimes are used. Additionally contains the fields:

stimtrain::Dict{String,T} – Dict in the format name/id => time. time should be relative to trigger event. Specified by the user on creation of struct using LaskaCore.relativespikes.

specs::Dict{String,T} – Dict containing the time before/after (back/forward) trigger events that spikes are included; as well as number of trigger events (ntrig).

Interfaces

Interfaces only for use with LaskaCore.RelativeSpikes.

LaskaCore.relativespecs — Function

relativespecs(rel::RelativeSpikes{T}) 
relativespecs(rel::RelativeSpikes{T}, spec::String) where {T<:Real}

Returns a Dict containing the 'specs' of a RelativeSpikes struct. Includes the back and forward variables used as well as the number of trigger events (ntrigs).

LaskaCore.stimtimes — Function

stimtimes(experiment::RelativeSpikes)

Returns a dict containing the stimtrain of a RelativeSpikes struct in the form of label => time.

Single cluster wrappers

Cluster

LaskaCore.Cluster — Type

struct Cluster{T} <: AbstractCluster{T}
    id::Int64
    info::Dict{String,String}
    spiketimes::Vector{T}
end

Struct for holding a single Cluster.

Direct field access is not recommended. Basic interface functions include:

LaskaCore.id – Returns the Cluster id.

LaskaCore.nspikes – Returns the number of spikes in the cluster (Based off length of the spiketimes field).

LaskaCore.info – Returns the info of the Cluster from "cluster_info.tsv" as a Dict.

LaskaCore.spiketimes – Returns a Vector containing all spiketimes of the Cluster.

RelativeCluster

LaskaCore.RelativeCluster — Type

struct RelativeCluster{T} <: AbstractCluster{T}
    id::Int64
    info::Dict{String,String}
    spiketimes::Vector{Vector{T}}
end

Struct for holding a cluster and its spiketimes relative to triggers. Similar to Cluster{T} except that the field spiketimes is a Vector{Vector{T}} where each vector represents trigger #n.

Direct field access is not recommended. Basic interface functions include:

LaskaCore.id – Returns the Cluster id.

LaskaCore.nspikes – Returns the number of spikes in the cluster (Based off length of the spiketimes field).

LaskaCore.info – Returns the info of the Cluster from "cluster_info.tsv" as a Dict.

LaskaCore.spiketimes – Returns a Vector containing all spiketimes of the Cluster.